LayoutTests/svg/filters/feComponentTransfer-style-crash-expected.txt

ALERT: TypeError: Type error
PASS
